requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
python-dotenv==1.0.0
requests==2.31.0
Flask-Login==0.6.3

# Security dependencies
Flask-WTF==1.2.1
bleach==6.1.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
python-magic-bin==0.4.14

# Image processing
Pillow==12.1.0
